# api/openapi.yaml
openapi: 3.1.0
info:
  title: Katalog API
  description: >
    REST API сервиса Katalog. Источник правды для контрактов:
    клиентские типы генерируются из этого файла (orval).
    Аутентификация: сессионная httpOnly cookie `katalog_session`.
    Тенант-изоляция: любой чужой или несуществующий ресурс — 404.
  version: 0.5.0
servers:
  - url: /
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Регистрация по email и паролю (argon2id), создаёт сессию
      requestBody:
        $ref: "#/components/requestBodies/Credentials"
      responses:
        "201":
          description: Пользователь создан, cookie установлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Вход, создаёт сессию
      requestBody:
        $ref: "#/components/requestBodies/Credentials"
      responses:
        "200":
          description: Сессия создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Выход (инвалидация сессии)
      responses:
        "204": { description: Сессия удалена }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/me:
    get:
      operationId: me
      summary: Текущий пользователь
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/password/forgot:
    post:
      operationId: forgotPassword
      summary: Запросить сброс пароля (письмо со ссылкой)
      description: >
        Всегда 204 — существование email не раскрывается. Если email
        зарегистрирован, на него уходит одноразовая ссылка (TTL 1 час).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: Принято }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/password/reset:
    post:
      operationId: resetPassword
      summary: Установить новый пароль по токену из письма
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "204": { description: Пароль обновлён }
        "400":
          description: Токен недействителен/использован (error=invalid_token) или слабый пароль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Подтвердить email по токену из письма регистрации
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "204": { description: Email подтверждён }
        "400":
          description: Токен недействителен или использован (error=invalid_token)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/shops:
    post:
      operationId: createShop
      summary: Создать магазин (slug валидируется, системные слова запрещены)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug: { type: string, minLength: 3, maxLength: 63 }
                name: { type: string, maxLength: 200 }
                description: { type: string }
                contacts: { type: object }
      responses:
        "201":
          description: Магазин создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      operationId: listShops
      summary: Магазины текущего пользователя
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/shops/{shopID}:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    get:
      operationId: getShop
      summary: Магазин по id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateShop
      summary: Обновить магазин (частично)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 200 }
                description: { type: string }
                contacts: { type: object }
                settings: { type: object }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteShop
      summary: Удалить магазин
      responses:
        "204": { description: Удалено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/shops/{shopID}/billing:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    get:
      operationId: getBilling
      summary: Тариф, лимиты, использование и подписка магазина
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Billing"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/shops/{shopID}/billing/subscribe:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    post:
      operationId: subscribe
      summary: Начать оплату платного тарифа через ЮKassa
      description: >
        Создаёт платёж ЮKassa (с сохранением способа оплаты для рекуррентных
        списаний) и возвращает confirmation_url для redirect. Тариф
        активируется только после вебхука payment.succeeded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan: { type: string, enum: [basic, pro] }
      responses:
        "200":
          description: Платёж создан
          content:
            application/json:
              schema:
                type: object
                required: [payment_id, confirmation_url]
                properties:
                  payment_id: { type: string, format: uuid }
                  confirmation_url: { type: string, format: uri }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502":
          description: ЮKassa недоступна (error=payment_provider_error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Платежи не сконфигурированы (error=billing_unavailable)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/shops/{shopID}/billing/cancel:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    post:
      operationId: cancelSubscription
      summary: Отключить автопродление подписки
      description: >
        Тариф действует до конца оплаченного периода, затем магазин проходит
        grace period и скрытие витрины. Идемпотентен.
      responses:
        "204": { description: Автопродление отключено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/billing/webhooks/yookassa:
    post:
      operationId: yooKassaWebhook
      summary: Вебхук ЮKassa о статусах платежей
      description: >
        Без сессии. Уведомления ЮKassa не подписаны, поэтому статус платежа
        перепроверяется прямым запросом к API ЮKassa. Обработка идемпотентна:
        повторная доставка события — no-op с 200. Не-2xx ответ означает,
        что ЮKassa должна доставить уведомление повторно.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, event, object]
              properties:
                type: { type: string, const: notification }
                event:
                  type: string
                  enum: [payment.succeeded, payment.canceled, payment.waiting_for_capture]
                object:
                  type: object
                  description: Объект платежа ЮKassa
      responses:
        "200":
          description: Уведомление обработано (или уже было обработано)
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Платёж не зарегистрирован (error=unknown_payment), нужен ретрай
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Не удалось перепроверить платёж в ЮKassa, нужен ретрай
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Платежи не сконфигурированы
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/shops/{shopID}/albums:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    post:
      operationId: createAlbum
      summary: Создать альбом (вложенность максимум 2 уровня)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 200 }
                parent_id: { type: string, format: uuid }
                sort_order: { type: integer }
      responses:
        "201":
          description: Альбом создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      operationId: listAlbums
      summary: Альбомы магазина
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/shops/{shopID}/albums/{albumID}:
    parameters:
      - $ref: "#/components/parameters/ShopID"
      - $ref: "#/components/parameters/AlbumID"
    get:
      operationId: getAlbum
      summary: Альбом по id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateAlbum
      summary: Обновить альбом (title, sort_order, is_hidden, cover_photo_id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
                sort_order: { type: integer }
                is_hidden: { type: boolean }
                cover_photo_id:
                  type: [string, "null"]
                  format: uuid
                  description: Пустая строка сбрасывает обложку; фото должно принадлежать альбому
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteAlbum
      summary: Удалить альбом
      responses:
        "204": { description: Удалено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/shops/{shopID}/albums/{albumID}/photos:
    parameters:
      - $ref: "#/components/parameters/ShopID"
      - $ref: "#/components/parameters/AlbumID"
    get:
      operationId: listPhotos
      summary: Фото альбома (со статусами обработки)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/uploads/presign:
    post:
      operationId: presignUpload
      summary: Pre-signed URL для прямой загрузки оригинала в S3
      description: >
        Проверяет владение альбомом и квоту тарифа. Создаёт запись фото
        в статусе uploading. Файл загружается PUT-запросом напрямую в S3.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shop_id, album_id, size]
              properties:
                shop_id: { type: string, format: uuid }
                album_id: { type: string, format: uuid }
                size: { type: integer, minimum: 1, maximum: 52428800 }
      responses:
        "200":
          description: URL выдан
          content:
            application/json:
              schema:
                type: object
                required: [photo_id, url]
                properties:
                  photo_id: { type: string, format: uuid }
                  url: { type: string, format: uri }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: >
            Мягкий отказ: квота тарифа исчерпана (error=quota_exceeded —
            хранилище, error=photo_quota_exceeded — число фото) или подписка
            неактивна (error=subscription_inactive, состояние grace/suspended).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/photos/confirm:
    post:
      operationId: confirmPhotos
      summary: Подтвердить загрузку — постановка в очередь обработки
      description: >
        Для каждого фото: проверка объекта в S3, uploading -> processing,
        задача воркеру (asynq). Ответ по каждому фото отдельно.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shop_id, photo_ids]
              properties:
                shop_id: { type: string, format: uuid }
                photo_ids:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: Результаты по каждому фото
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [photo_id, status]
                      properties:
                        photo_id: { type: string }
                        status: { type: string }
                        error: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/photos/{photoID}:
    parameters:
      - name: photoID
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updatePhoto
      summary: Обновить подпись фото
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption: { type: string, maxLength: 2000 }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deletePhoto
      summary: Удалить фото (объекты S3, квота, счётчик альбома)
      responses:
        "204": { description: Удалено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---- Публичный неймспейс витрины (без сессии, отдельные схемы ответов:
  # ---- приватные поля наружу не отдаются никогда) ----
  /api/v1/public/shops/{slug}:
    get:
      operationId: getPublicShop
      summary: Шапка магазина и сетка не скрытых альбомов
      parameters:
        - $ref: "#/components/parameters/ShopSlug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [shop, albums]
                properties:
                  shop: { $ref: "#/components/schemas/PublicShop" }
                  albums:
                    type: array
                    items: { $ref: "#/components/schemas/PublicAlbum" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/shops/{slug}/albums/{albumID}:
    get:
      operationId: getPublicAlbum
      summary: Фото альбома (только ready) с пагинацией
      parameters:
        - $ref: "#/components/parameters/ShopSlug"
        - $ref: "#/components/parameters/AlbumID"
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 60 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [shop, album, photos, page, per_page, total]
                properties:
                  shop: { $ref: "#/components/schemas/PublicShop" }
                  album: { $ref: "#/components/schemas/PublicAlbum" }
                  photos:
                    type: array
                    items: { $ref: "#/components/schemas/PublicPhoto" }
                  page: { type: integer }
                  per_page: { type: integer }
                  total: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/shops/{slug}/search:
    get:
      operationId: searchPublicPhotos
      summary: Поиск по подписям в рамках магазина (FTS russian + pg_trgm fallback)
      parameters:
        - $ref: "#/components/parameters/ShopSlug"
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: OK (максимум 50 результатов)
          content:
            application/json:
              schema:
                type: object
                required: [query, photos]
                properties:
                  query: { type: string }
                  photos:
                    type: array
                    items: { $ref: "#/components/schemas/PublicPhoto" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/lead-click:
    post:
      operationId: createLeadClick
      summary: >
        Фиксация клика «написать по товару». visitor_hash строится из
        анонимной cookie kv (устанавливается сервером) и IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shop_id, channel]
              properties:
                shop_id: { type: string, format: uuid }
                photo_id: { type: string, format: uuid }
                channel: { type: string, enum: [telegram, whatsapp, max, vk] }
      responses:
        "204": { description: Лид зафиксирован }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/sitemap:
    get:
      operationId: getPublicSitemap
      summary: Активные магазины для sitemap витрины
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [shops]
                properties:
                  shops:
                    type: array
                    items:
                      type: object
                      required: [slug, updated_at]
                      properties:
                        slug: { type: string }
                        updated_at: { type: string, format: date-time }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/complaints:
    post:
      operationId: createComplaint
      summary: Жалоба правообладателя на контент (notice-and-takedown), без auth
      description: >
        Жалоба привязывается к магазину/фото, если ссылку удалось распознать
        (/{slug} или /media/{shop_id}/{photo_id}/...); нераспознанный URL
        не мешает подать жалобу. Модератор получает email-уведомление.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, reporter_name, reporter_email, reason]
              properties:
                url: { type: string, maxLength: 1000 }
                reporter_name: { type: string, maxLength: 200 }
                reporter_email: { type: string, format: email }
                reason: { type: string, minLength: 10, maxLength: 5000 }
      responses:
        "201":
          description: Жалоба зарегистрирована
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string, const: open }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/admin/complaints:
    get:
      operationId: adminListComplaints
      summary: "[admin] Список жалоб (фильтр по статусу)"
      description: Все /api/v1/admin/* доступны только role=admin; остальным — 404.
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [open, in_review, resolved, rejected] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminComplaint" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/complaints/{complaintID}:
    patch:
      operationId: adminUpdateComplaint
      summary: "[admin] Сменить статус жалобы (с записью в аудит-лог)"
      parameters:
        - name: complaintID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, in_review, resolved, rejected] }
      responses:
        "200":
          description: Статус обновлён
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/photos/flagged:
    get:
      operationId: adminListFlaggedPhotos
      summary: "[admin] Фото, помеченные стоп-словами на ручную проверку"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, shop_id, shop_slug, album_id, caption, status]
                  properties:
                    id: { type: string, format: uuid }
                    shop_id: { type: string, format: uuid }
                    shop_slug: { type: string }
                    album_id: { type: string, format: uuid }
                    caption: { type: string }
                    status: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/admin/photos/{photoID}/unflag:
    post:
      operationId: adminUnflagPhoto
      summary: "[admin] Снять флаг ручной проверки (ложное срабатывание стоп-слова)"
      parameters:
        - name: photoID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Флаг снят }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/albums/{albumID}/hide:
    post:
      operationId: adminHideAlbum
      summary: "[admin] Скрыть альбом с витрины (is_hidden), email владельцу"
      parameters:
        - name: albumID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminActionBody" }
      responses:
        "204": { description: Альбом скрыт }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  requestBodies:
    Credentials:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [email, password]
            properties:
              email: { type: string, format: email }
              password: { type: string, minLength: 8 }
  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Нет или истекла сессия / неверные креды
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: Конфликт (занятый email/slug)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Превышен rate-limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Машинный код ошибки }
        message: { type: string }
    User:
      type: object
      required: [id, role, email_verified]
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        role: { type: string, enum: [user, admin] }
        email_verified: { type: boolean }
...